Add config param to use api without db
